rename stables balance card
